requirements.txt
discord.py-self[voice]
flask
redis
pillow
requests